# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Hunt Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: Get current user's bookmarks
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"

    post:
      operationId: addBookmark
      tags: [bookmarks]
      summary: Add a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmarked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{toolId}:
    delete:
      operationId: removeBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      parameters:
        - name: toolId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
  /bookmarks/check/{toolId}:
    get:
      operationId: checkBookmark
      tags: [bookmarks]
      summary: Check if a tool is bookmarked
      parameters:
        - name: toolId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Bookmark status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkStatus"
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBlogPosts"

    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
  /blog/id/{id}:
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Subscribed
...
  /admin/tools:
    get:
      operationId: adminListTools
      tags: [admin]
      summary: Admin list all tools with any status
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected, featured] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Tools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTools"
  /admin/tools/{id}/status:
    patch:
      operationId: updateToolStatus
      tags: [admin]
      summary: Update tool status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolStatusUpdate"
      responses:
        "200":
          description: Updated tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin list users
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /admin/reports:
    get:
      operationId: adminListReports
      tags: [admin]
      summary: Admin list reports
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, resolved, dismissed] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReports"
  /admin/reports/{id}/status:
    patch:
      operationId: updateReportStatus
      tags: [admin]
      summary: Update report status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportStatusUpdate"
      responses:
        "200":
          description: Updated report
  /admin/newsletter:
    get:
      operationId: adminListSubscribers
      tags: [admin]
      summary: Admin list newsletter subscribers
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Subscribers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSubscribers"
components:
  schemas:
...
    Tool:
      type: object
      required: [id, name, slug, websiteUrl, shortDescription, pricing, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        websiteUrl: { type: string }
        shortDescription: { type: string }
        longDescription: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        demoVideoUrl: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        categorySlug: { type: ["string", "null"] }
        pricing: { type: string, enum: [free, freemium, paid, contact] }
        status: { type: string, enum: [pending, approved, rejected, featured] }
        isVerified: { type: boolean }
        isEditorChoice: { type: boolean }
        views: { type: integer }
        clicks: { type: integer }
        bookmarkCount: { type: integer }
        ratingAvg: { type: number }
        reviewCount: { type: integer }
        ownerName: { type: ["string", "null"] }
        contactEmail: { type: ["string", "null"] }
        supportUrl: { type: ["string", "null"] }
        documentationUrl: { type: ["string", "null"] }
        githubUrl: { type: ["string", "null"] }
        launchDate: { type: ["string", "null"] }
        aiModel: { type: ["string", "null"] }
        features: { type: array, items: { type: string } }
        pros: { type: array, items: { type: string } }
        cons: { type: array, items: { type: string } }
        platforms: { type: array, items: { type: string } }
        integrations: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        screenshots: { type: array, items: { type: string } }
        socialLinks: { type: object, additionalProperties: { type: string } }
        submittedBy: { type: ["string", "null"] }
        trendingScore: { type: number }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    ToolStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, approved, rejected, featured] }
        isVerified: { type: boolean }
        isEditorChoice: { type: boolean }
    PaginatedTools:
      type: object
      required: [tools, total, page, limit]
      properties:
        tools:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    Bookmark:
      type: object
      required: [id, toolId, createdAt]
      properties:
        id: { type: integer }
        toolId: { type: integer }
        userId: { type: string }
        tool:
          $ref: "#/components/schemas/Tool"
        createdAt: { type: string }
    BookmarkInput:
      type: object
      required: [toolId]
      properties:
        toolId: { type: integer }
    BookmarkStatus:
      type: object
      required: [bookmarked]
      properties:
        bookmarked: { type: boolean }
    BlogPost:
      type: object
      required: [id, title, slug, excerpt, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: ["string", "null"] }
        authorName: { type: ["string", "null"] }
        authorAvatar: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        readingTime: { type: integer }
        coverImage: { type: ["string", "null"] }
        status: { type: string, enum: [draft, published] }
        publishedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    BlogPostInput:
      type: object
      required: [title, slug, excerpt]
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        authorName: { type: string }
        authorAvatar: { type: string }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        coverImage: { type: string }
        status: { type: string, enum: [draft, published] }
    BlogPostUpdate:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        authorName: { type: string }
        authorAvatar: { type: string }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        coverImage: { type: string }
        status: { type: string, enum: [draft, published] }
    PaginatedBlogPosts:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    NewsletterInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
...
    UserProfile:
      type: object
      required: [id, clerkUserId, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        email: { type: ["string", "null"] }
        name: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        role: { type: string, enum: [user, admin] }
        createdAt: { type: string }
...
    PaginatedUsers:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    Report:
      type: object
      required: [id, toolId, reason, status, createdAt]
      properties:
        id: { type: integer }
        toolId: { type: integer }
        reporterId: { type: ["string", "null"] }
        reason: { type: string }
        details: { type: ["string", "null"] }
        status: { type: string, enum: [open, resolved, dismissed] }
        createdAt: { type: string }
        tool:
          $ref: "#/components/schemas/Tool"
    ReportStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [open, resolved, dismissed] }
    PaginatedReports:
      type: object
      required: [reports, total, page, limit]
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    NewsletterSubscriber:
      type: object
      required: [id, email, subscribedAt]
      properties:
        id: { type: integer }
        email: { type: string }
        subscribedAt: { type: string }
    PaginatedSubscribers:
      type: object
      required: [subscribers, total, page, limit]
      properties:
        subscribers:
          type: array
          items:
            $ref: "#/components/schemas/NewsletterSubscriber"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }