# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Hunt Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Tool } from './tool';

export interface Bookmark {
  id: number;
  toolId: number;
  userId?: string;
  tool?: Tool;
  createdAt: string;
}